fix: split TermWidth into platform files for Windows cross-compilation
Move TermWidth() and the TIOCGWINSZ ioctl logic from ansi.go into
ansi_unix.go (behind //go:build unix). Add ansi_windows.go with a
COLUMNS env fallback for Windows. This fixes goreleaser builds that
fail on windows/amd64 due to syscall.SYS_IOCTL being undefined.
